obsolete test facts names
