+++ src/Orc.Automation.Tests/TestFacts/Base/AppUiTestsBase.cs
@@ -4,7 +4,7 @@ using System.Windows.Automation;
 using Controls;
 using NUnit.Framework;
 
-public abstract class AppUiTestsBase<TMainWindow> : UiTestFactsBase
+public abstract class AppUiTestsBase<TMainWindow> : UiTestsBase
     where TMainWindow : class
 {
     #region Properties
